added confetti effect if the player wins, organized the project, and did some overall polishing.
