+++ tictak/Features/Game/Domain/Difficulty.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+enum Difficulty: String, CaseIterable, Identifiable {
+    case easy = "Easy"
+    case medium = "Medium"
+    case hard = "Hard"
+    
+    var id: String { rawValue }
+}
