+++ tictak/Features/Game/Domain/Move.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+struct Move {
+    let player: Player
+    let boardIndex: Int
+    
+    var indicator: String {
+        player == .human ? "xmark" : "circle"
+    }
+}
