+++ tictak/Features/Game/Domain/Player.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+enum Player {
+    case human, computer
+}
